% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedicalRiskPredictionModels.R
\docType{data}
\name{as}
\alias{as}
\title{as}
\format{
A data frame with 217 observations on the following 10 variables.
\describe{
\item{age5}{Age in years divided by 5}
\item{psa}{psa}
\item{ct1}{ct1}
\item{diaggs}{diaggs}
\item{ppb5}{ppb5}
\item{lmax}{lmax}
\item{asprog}{Progression status: 1= progression, 2= death without progression, 0 = censored}
\item{asprogtime}{Time on active surveillance}
\item{age}{Age in years}
\item{erg}{Biomarker erg status}
}
}
\source{
The data are computer modified synthetic data
}
\description{
This is a computer modified synthetic version of the data of the active 
surveillance study by Berg et al.
}
\examples{
data(as)
}
\references{
Kasper Drimer Berg, Ben Vainer, Frederik Birkebaek Thomsen, M Andreas
Roeder, Thomas Alexander Gerds, Birgitte Groenkaer Toft, Klaus Brasso, and
Peter Iversen. Erg protein expression in diagnostic specimens is associated
with increased risk of progression during active surveillance for prostate
cancer. European urology, 66(5):851--860, 2014.
}
\keyword{datasets}
