% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMRPM.R
\name{dataMRPM}
\alias{dataMRPM}
\title{load data for Medical Risk Prediction Models}
\usage{
dataMRPM()
}
\value{
4 data sets
}
\description{
Loads 4 data sets to illustrate the R-code of the book Medical Risk Prediction Models by Gerds & Kattan
}
\details{
The data are synthetic data generated with the computer.
}
\examples{
dataMRPM()
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
