% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedicalRiskPredictionModels.R
\docType{data}
\name{ivf}
\alias{ivf}
\title{ivf}
\format{
A data frame with 276 observations on the following 11 variables.
\describe{
\item{\code{cyclelen}}{Cycle length} 
\item{\code{bmi}}{BMI}
\item{\code{weight}}{Body weight}
\item{\code{age}}{Age}
\item{\code{ant.foll}}{Number of antral follicles}
\item{\code{fsh}}{Hormon fsh}
\item{\code{smoking}}{Smoking status}
\item{\code{no.cig.d}}{Number of cigarettes per day}
\item{\code{ovolume}}{Ovarian volume}
\item{\code{ohss}}{Ovarian hyper stimulation syndrome: as integer}
\item{\code{ohss}}{Ovarian hyper stimulation syndrome: as character}
}
}
\source{
The data are computer modified synthetic data
}
\description{
This is a computer modified synthetic version of the data of the in
vitro fertilization study by la Cour Freiesleben et al.
}
\examples{
data(ivf)
}
\references{
N. la Cour Freiesleben, T.A. Gerds, J.L. Forman, J.D. Silver, A. Nyboe
Andersen, and B. Popovic-Todorovic. Risk charts to identify low and
excessive responders among first-cycle ivf/icsi standard patients.
Reproductive Biomedicine Online, 22(1):50--58, 2011.
}
\keyword{datasets}
