% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedicalRiskPredictionModels.R
\docType{data}
\name{long}
\alias{long}
\title{long}
\format{
A data frame with 18 observations on the following 3 variables.
\describe{
\item{subject}{subject}
\item{psadate}{psadate}
\item{psa}{psa}
}
}
\source{
The data are computer modified synthetic data
}
\description{
This is a computer generated data set for illustration purposes
}
\examples{
data(long)
}
\keyword{datasets}
