% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedicalRiskPredictionModels.R
\docType{data}
\name{oc}
\alias{oc}
\title{oc}
\format{
A data frame with 1284 observations on the following 14 variables.
\describe{
\item{\code{survtime}}{survival time in months} 
\item{\code{survstatus}}{survival status}
\item{\code{grade}}{Tumor grade}
\item{\code{age}}{Age}
\item{\code{tumorthickness}}{Tumor thickness}
\item{gender}{gender}
\item{tobacco}{tobacco}
\item{deep}{deep}
\item{site}{site}
\item{race}{race}
\item{x.posnodes}{number of positive lymphnodes}
\item{tumormaxdimension}{tumormaxdimension}
\item{vascular}{vascular}
\item{Grade}{Grade}
}
}
\source{
The data are computer modified synthetic data
}
\description{
This is a computer modified synthetic version of the data of the oral 
cancer study by
}
\examples{
data(oc)
}
\references{

}
\keyword{datasets}
