% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareExamples.R
\name{prepareExamples}
\alias{prepareExamples}
\title{load data for Medical Risk Prediction Models}
\usage{
prepareExamples()
}
\value{
modified global environment with data and libraries loaded
}
\description{
Loads libraries and data sets to illustrate the R-code of
the book Medical Risk Prediction Models by Gerds & Kattan
}
\details{
These are synthetic data generated with the computer.
}
\examples{
prepareExamples()
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
