% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedicalRiskPredictionModels.R
\docType{data}
\name{ttedata}
\alias{ttedata}
\title{ttedata}
\format{
A data frame with 5 observations on the following 5 variables.
\describe{
\item{id}{id}
\item{af.date}{af date}
\item{death.date}{death date}
\item{stroke.date}{stroke date}
\item{lost.date}{end of followup date}
}
}
\source{
The data are computer modified synthetic data
}
\description{
This is a computer generated data set for illustration purposes
}
\examples{
data(ttedata)
}
\keyword{datasets}
